File Save: add extension by default
Previously, an extension was automatically added after save if none
had been entered in the filename. However, this affected file overwrite
detection, as Gtk wouldn't show the overwrite confirmation dialog if
file "subtitles" was entered and a "subtitles.srt" file already existed
(this would be the final filename). So now we automatically set the
extension when the dialog is opened and when changing between subtitle
formats.