Fix #4 - Syslog full of "source ID not found"
This problem is actually in gtk-sharp. Ignoring this message until
there's a gtk-sharp release fixing this problem (note: the fix has
actually been imlpemented - see the commit diff for details - but not
release yet; should come with 2.99.4).

Complete message:
Source ID <number> was not found when attempting to remove it