Fix #153: Add support for Enchant 2
The API we use hasn't changed from ver. 1, however the config file was
pointing specifically to ver. 1. With this change, the config file isn't
used anymore, instead we dynamically load libenchant according to the
following order: ver2, ver1, "any version" (libenchant.so, as a fallback
method if the other two don't exist).